# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EdgeFinder sports betting value prediction API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /games:
    get:
      operationId: getGames
      tags: [games]
      summary: Get live and upcoming games
      parameters:
        - name: sport
          in: query
          schema:
            type: string
            enum: [NFL, NBA, MLB, NHL, all]
            default: all
        - name: status
          in: query
          schema:
            type: string
            enum: [live, upcoming, final, all]
            default: all
      responses:
        "200":
          description: List of games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Game"
  /games/{gameId}:
    get:
      operationId: getGame
      tags: [games]
      summary: Get a specific game with odds and prediction
      parameters:
        - name: gameId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Game detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /predictions:
    get:
      operationId: getPredictions
      tags: [predictions]
      summary: Get model predictions for games
      parameters:
        - name: sport
          in: query
          schema:
            type: string
            enum: [NFL, NBA, MLB, NHL, all]
            default: all
        - name: minEdge
          in: query
          schema:
            type: number
            default: 0
      responses:
        "200":
          description: List of predictions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Prediction"
  /value-bets:
    get:
      operationId: getValueBets
      tags: [predictions]
      summary: Get value bets with 3%+ edge flagged
      parameters:
        - name: sport
          in: query
          schema:
            type: string
            enum: [NFL, NBA, MLB, NHL, all]
            default: all
        - name: minEdge
          in: query
          schema:
            type: number
            default: 3
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [edge, odds, kellyStake, sport]
            default: edge
        - name: sortDir
          in: query
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        "200":
          description: Value bets sorted by edge
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ValueBet"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required: [error]
...
    Game:
      type: object
      properties:
        id:
          type: string
        sport:
          type: string
          enum: [NFL, NBA, MLB, NHL]
        homeTeam:
          type: string
        awayTeam:
          type: string
        homeScore:
          type: integer
          nullable: true
        awayScore:
          type: integer
          nullable: true
        startTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [live, upcoming, final]
        quarter:
          type: string
          nullable: true
        timeRemaining:
          type: string
          nullable: true
        venue:
          type: string
          nullable: true
        weather:
          $ref: "#/components/schemas/WeatherInfo"
          nullable: true
        topEdge:
          type: number
          nullable: true
      required: [id, sport, homeTeam, awayTeam, startTime, status]
    WeatherInfo:
      type: object
      properties:
        temp:
          type: number
        condition:
          type: string
        windSpeed:
          type: number
        precipitation:
          type: number
      required: [temp, condition, windSpeed, precipitation]
    GameDetail:
      type: object
      properties:
        game:
          $ref: "#/components/schemas/Game"
        odds:
          type: array
          items:
            $ref: "#/components/schemas/OddsEntry"
        prediction:
          $ref: "#/components/schemas/Prediction"
          nullable: true
        injuries:
          type: array
          items:
            $ref: "#/components/schemas/InjuryReport"
      required: [game, odds]
    InjuryReport:
      type: object
      properties:
        player:
          type: string
        team:
          type: string
        status:
          type: string
        position:
          type: string
        impact:
          type: string
          enum: [high, medium, low]
      required: [player, team, status, position, impact]
    OddsEntry:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        bookmaker:
          type: string
        homeMoneyline:
          type: integer
          nullable: true
        awayMoneyline:
          type: integer
          nullable: true
        homeSpread:
          type: number
          nullable: true
        awaySpread:
          type: number
          nullable: true
        homeSpreadOdds:
          type: integer
          nullable: true
        awaySpreadOdds:
          type: integer
          nullable: true
        overUnder:
          type: number
          nullable: true
        overOdds:
          type: integer
          nullable: true
        underOdds:
          type: integer
          nullable: true
        updatedAt:
          type: string
          format: date-time
      required: [id, gameId, bookmaker, updatedAt]
    Prediction:
      type: object
      properties:
        gameId:
          type: string
        sport:
          type: string
        homeWinProb:
          type: number
        awayWinProb:
          type: number
        drawProb:
          type: number
          nullable: true
        edgeHome:
          type: number
        edgeAway:
          type: number
        recommendedBet:
          type: string
          nullable: true
        recommendedBookmaker:
          type: string
          nullable: true
        confidence:
          type: string
          enum: [high, medium, low]
        modelFactors:
          type: array
          items:
            type: string
        updatedAt:
          type: string
          format: date-time
      required: [gameId, sport, homeWinProb, awayWinProb, edgeHome, edgeAway, confidence, modelFactors, updatedAt]
    ValueBet:
      type: object
      properties:
        id:
          type: string
        gameId:
          type: string
        sport:
          type: string
          enum: [NFL, NBA, MLB, NHL]
        homeTeam:
          type: string
        awayTeam:
          type: string
        startTime:
          type: string
          format: date-time
        team:
          type: string
        betType:
          type: string
          enum: [moneyline, spread, over, under]
        bookmaker:
          type: string
        odds:
          type: integer
        impliedProb:
          type: number
        modelProb:
          type: number
        edge:
          type: number
        kellyStake:
          type: number
        status:
          type: string
          enum: [live, upcoming]
      required: [id, gameId, sport, homeTeam, awayTeam, startTime, team, betType, bookmaker, odds, impliedProb, modelProb, edge, kellyStake, status]
...