# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeFinder sports betting value prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { GameSport } from "./gameSport";
import type { GameStatus } from "./gameStatus";
import type { WeatherInfo } from "./weatherInfo";

export interface Game {
  id: string;
  sport: GameSport;
  homeTeam: string;
  awayTeam: string;
  homeScore?: number | null;
  awayScore?: number | null;
  startTime: Date;
  status: GameStatus;
  quarter?: string | null;
  timeRemaining?: string | null;
  venue?: string | null;
  weather?: WeatherInfo | null;
  topEdge?: number | null;
}