# lib/api-zod/src/generated/types/weatherInfo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * EdgeFinder sports betting value prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface WeatherInfo {
  temp: number;
  condition: string;
  windSpeed: number;
  precipitation: number;
}